% Auto-generated: do not edit by hand
\name{dashRotaryKnob}

\alias{dashRotaryKnob}

\title{DashRotaryKnob component}

\description{

}

\usage{
dashRotaryKnob(id=NULL, className=NULL, format=NULL, max=NULL, min=NULL,
preciseMode=NULL, skinName=NULL, step=NULL, style=NULL,
unlockDistance=NULL, value=NULL)
}

\arguments{
\item{id}{Character. }

\item{className}{Character. }

\item{format}{Character. }

\item{max}{Numeric. }

\item{min}{Numeric. }

\item{preciseMode}{Logical. }

\item{skinName}{Character. }

\item{step}{Numeric. }

\item{style}{Named list. }

\item{unlockDistance}{Numeric. }

\item{value}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

